pyproject.toml
[tool.poetry]
name = "pytickersymbols"
version = "1.15.0"
description = "pytickersymbols provides access to google and yahoo ticker symbols"
authors = ["SlashGordon <slash.gordon.dev@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
wheel = ">=0.41.3,<0.43.0"
pytest = ">=7.4.3,<9.0.0"
pytest-cov = "^4.1.0"
PyYAML = "^6.0.1"
pycountry = ">=22.3.5,<24.0.0"
packaging = "^23.0"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
click = "^8.3.1"
yfinance = "^0.2.66"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "src/pytickersymbols/__init__.py:__version__",
]
version_toml = [
    "pyproject.toml:tool.poetry.version",
]
branch = "master"
upload_to_pypi = true
upload_to_release = true
build_command = "source .venv/bin/activate && python3 tools/yaml2data.py && poetry build"

[tool.pytest.ini_options]
addopts = "tests/ --junitxml test-results/test.xml --cov src/pytickersymbols --cov-report term-missing"
pythonpath = ["src"]